Separate functions for applying type annotations
